make the navigation bar work as a server component this allows the 404 to have a nav as well as a single nav bar that works with i18n
