+++ apps/web/src/app/[locale]/layout.tsx
@@ -1,5 +1,4 @@
 'use client';
-import { Navigation } from '~/components/navigation';
 import { I18nProviderClient } from '~/locales/client';
 
 export default function RootLayout({
@@ -12,7 +11,6 @@ export default function RootLayout({
   return (
     <>
       <I18nProviderClient locale={locale}>
-        <Navigation />
         <main className="flex-1">{children}</main>
       </I18nProviderClient>
     </>
