+++ apps/web/src/app/layout.tsx
@@ -67,6 +67,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
       </head>
       <body className={`${inter.className} flex flex-col`}>
         <Providers>
+          <Navigation />
           {children}
           <Toaster />
         </Providers>
