+++ apps/web/src/components/theme-button.tsx
@@ -1,0 +1,31 @@
+'use client';
+import { useEffect, useState } from 'react';
+import { useTheme } from 'next-themes';
+import { Moon, Sun } from '@repo/ui/icons';
+
+export function ThemeButton() {
+  const [mounted, setMounted] = useState(false);
+  const { setTheme, resolvedTheme } = useTheme();
+
+  useEffect(() => {
+    setMounted(true);
+  }, []);
+
+  return (
+    <>
+      {mounted ? (
+        <button
+          aria-label="theme button"
+          className="focus:bg-accent rounded-lg p-2 duration-300 focus:outline-none focus-visible:ring-2"
+          onClick={() => {
+            setTheme(resolvedTheme === 'dark' ? 'light' : 'dark');
+          }}
+          type="button"
+        >
+          {resolvedTheme === 'dark' && <Moon aria-hidden="true" className="h-5 w-5" />}
+          {resolvedTheme === 'light' && <Sun aria-hidden="true" className="h-5 w-5" />}
+        </button>
+      ) : null}
+    </>
+  );
+}
